# openapi.yaml
openapi: 3.0.3
info:
  title: Tanabata File Manager API
  description: |
    REST API for Tanabata File Manager — a multi-user, tag-based web file manager.

    ## Authentication
    All endpoints except `POST /auth/login` require a Bearer JWT token
    in the `Authorization` header.

    ## Pagination
    - **Files**: cursor-based (`cursor` parameter, returned in `next_cursor`).
    - **All other lists**: offset-based (`offset` + `limit`).

    ## Error format
    ```json
    {
      "code": "string",
      "message": "string",
      "details": [{ "field": "string", "message": "string" }]
    }
    ```

    ## File filter DSL
    Files can be filtered via a DSL in the `filter` query parameter.
    Tokens are comma-separated inside braces: `{token1,token2,...}`.

    Operators: `(`, `)`, `&` (AND), `|` (OR), `!` (NOT).
    Conditions: `t=<tag_uuid>` (has tag), `t=00000000-0000-0000-0000-000000000000` (untagged),
    `m=<mime_id>` (exact MIME), `m~<pattern>` (MIME LIKE pattern, e.g. `m~image%`).

    Example: `{t=uuid1,&,!,t=uuid2}` → has tag1 AND NOT tag2.
  version: 1.0.0
  license:
    name: Proprietary
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /files:
    get:
      tags: [Files]
      summary: List files (cursor-based pagination)
      parameters:
        - name: cursor
          in: query
          description: Cursor from previous response's `next_cursor` or `prev_cursor`
          schema:
            type: string
        - name: direction
          in: query
          description: Pagination direction relative to cursor
          schema:
            type: string
            enum: [forward, backward]
            default: forward
        - name: anchor
          in: query
          description: |
            File UUID to anchor the listing around. The response will
            include this file and its neighbors. Use with `direction`
            and `limit` to load items before/after the anchor.
            Mutually exclusive with `cursor`.
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: sort
          in: query
          description: Sort field
          schema:
            type: string
            enum: [content_datetime, created, original_name, mime]
            default: created
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: filter
          in: query
          description: Filter DSL expression (see API description)
          schema:
            type: string
        - name: search
          in: query
          description: Search by original_name (substring match)
          schema:
            type: string
        - name: trash
          in: query
          description: If true, return only soft-deleted files (trash)
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Paginated file list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileCursorPage'

    post:
      tags: [Files]
      summary: Upload a new file
      description: |
        Upload a file via multipart form. If `content_datetime` is not provided,
        it is extracted from EXIF data (DateTimeOriginal). EXIF metadata is
        automatically extracted and stored in the `exif` field (immutable).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                content_datetime:
                  type: string
                  format: date-time
                notes:
                  type: string
                metadata:
                  type: string
                  description: JSON string of key-value pairs
                is_public:
                  type: boolean
                  default: false
                tag_ids:
                  type: string
                  description: Comma-separated tag UUIDs to assign
      responses:
        '201':
          description: File created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
        '400':
          $ref: '#/components/responses/ValidationError'
        '415':
          description: Unsupported MIME type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /files/{file_id}:
    get:
      tags: [Files]
      summary: Get file metadata
      parameters:
        - $ref: '#/components/parameters/file_id'
      responses:
        '200':
          description: File metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Files]
      summary: Update file metadata
      parameters:
        - $ref: '#/components/parameters/file_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileUpdate'
      responses:
        '200':
          description: Updated file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Files]
      summary: Soft-delete file (move to trash)
      parameters:
        - $ref: '#/components/parameters/file_id'
      responses:
        '204':
          description: Moved to trash
        '404':
          $ref: '#/components/responses/NotFound'
...
  /tags/{tag_id}/files:
    get:
      tags: [Tags, Files]
      summary: List files with this tag
      parameters:
        - $ref: '#/components/parameters/tag_id'
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: File list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileCursorPage'

  # --- Tag rules ---
...
  /pools/{pool_id}/files:
    get:
      tags: [Pools, Files]
      summary: List files in a pool (ordered by position)
      parameters:
        - $ref: '#/components/parameters/pool_id'
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: filter
          in: query
          description: Filter DSL (same syntax as /files)
          schema:
            type: string
      responses:
        '200':
          description: Ordered file list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoolFileCursorPage'

    post:
      tags: [Pools, Files]
      summary: Add files to a pool
      parameters:
        - $ref: '#/components/parameters/pool_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [file_ids]
              properties:
                file_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                position:
                  type: integer
                  description: Insert position for the first file (rest follow sequentially; appended at end if omitted)
      responses:
        '201':
          description: Files added to pool
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # -------------------------------------------------------------------------
  # Parameters
  # -------------------------------------------------------------------------
  parameters:
    file_id:
      name: file_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    tag_id:
      name: tag_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    pool_id:
      name: pool_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  # -------------------------------------------------------------------------
  # Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: not_found
        message:
          type: string
          example: File not found
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string

    # --- Auth ---
...
    File:
      type: object
      properties:
        id:
          type: string
          format: uuid
        original_name:
          type: string
          nullable: true
        mime_type:
          type: string
          example: image/jpeg
        mime_extension:
          type: string
          example: jpg
        content_datetime:
          type: string
          format: date-time
        notes:
          type: string
          nullable: true
        metadata:
          type: object
          nullable: true
        exif:
          type: object
          readOnly: true
          description: EXIF data extracted at upload time (immutable, empty object if no EXIF)
        phash:
          type: integer
          format: int64
          nullable: true
        creator_id:
          type: integer
        creator_name:
          type: string
        is_public:
          type: boolean
        is_deleted:
          type: boolean
        created_at:
          type: string
          format: date-time
          description: Extracted from UUID v7
    FileUpdate:
      type: object
      properties:
        original_name:
          type: string
        content_datetime:
          type: string
          format: date-time
        notes:
          type: string
        metadata:
          type: object
        is_public:
          type: boolean
    FileCursorPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/File'
        next_cursor:
          type: string
          nullable: true
          description: Cursor for loading next (forward) page; null if no more items ahead
        prev_cursor:
          type: string
          nullable: true
          description: Cursor for loading previous (backward) page; null if at the beginning

    # --- Tag ---
...
    PoolFile:
      type: object
      allOf:
        - $ref: '#/components/schemas/File'
        - type: object
          properties:
            position:
              type: integer
    PoolFileCursorPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PoolFile'
        next_cursor:
          type: string
          nullable: true

    # --- ACL ---
...